Package just client tools/drivers from the nuodb server package
Switch the client package to only bundling client tools/drivers which
are only otherwise available in the nuodb server package. Any
tool/driver/library which is available from an external package
manager (pypi, maven, etc.) should be retrieved from that package
manager instead of from this package.

This commit also introduces the concept of a "bundle". Multiple tarballs
may be produced now by the build (instead of just one), one tarball per
bundle. The internal structure of each package is the same as before,
but its contents are determined by its bundle. Available bundles
are listed in the Bundle enum in cliend/bundles.py. A Stage should be
labeled with a bundle. Only stages in bundle B will be included in the
package for bundle B.

| Bundle     | Contents                   | Package                                  |
+------------+----------------------------+------------------------------------------+
| c-driver   | C driver                   | nuodb-c-driver-20230626.lin-x64.tar.gz   |
| cpp-driver | C++ driver                 | nuodb-cpp-driver-20230626.lin-x64.tar.gz |
| tools      | nuosql, nuoloader, nuodump | nuodb-tools-20230626.lin-x64.tar.gz      |

The C driver is also included in the cpp-driver bundle as it is a dependency.